[Tab#746] Split the tab view into two seperate views.
